Went back to the old header. I think I will just have to go back to the drawing board and pick a different CSS framework or just not use Shoelace. Also tidying up buttons and forms
