+++ memory-cards/public/index.html
@@ -26,7 +26,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>React App</title>
+    <title>Memory Cards</title>
   </head>
   <body>
     <noscript>
